backend/requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9       # required for FastAPI file uploads

# ── Config ───────────────────────────────────────────────────────────────────
pydantic>=2.7.0
pyyaml>=6.0.1

# ── HTTP client (local LLM backend calls + streaming) ────────────────────────
httpx>=0.27.0

# ── Document text extraction ─────────────────────────────────────────────────
pypdf>=4.2.0                  # PDF text layer extraction
python-docx>=1.1.2            # DOCX
python-pptx>=0.6.23           # PPTX read + write
docx2txt>=0.8                 # fallback DOCX extractor

# ── OCR (optional, comment out if Tesseract is not installed) ────────────────
# pdf2image>=1.17.0
# pytesseract>=0.3.10

# ── RAG / Embeddings ──────────────────────────────────────────────────────────
sentence-transformers>=2.7.0  # local embeddings (all-MiniLM-L6-v2)

# ── Token counting (optional, falls back to char/4 heuristic) ────────────────
# tiktoken>=0.7.0

# ── Cloud LLM providers (install only what you need) ─────────────────────────
openai>=1.30.0                 # OpenAI API – also works with LocalAI if needed
anthropic>=0.25.0              # Anthropic Claude API

# ── Database ─────────────────────────────────────────────────────────────────
# sqlite3 is part of the Python standard library – no extra package needed.
# Uncomment the line below if you prefer DuckDB:
# duckdb>=0.10.0
